fix(task-board): keep estimates frozen after start
